Docs changes

+++ cpp14/tests/src/tests Data Structure/tests Dynamic Array.cpp
@@ -9,12 +9,10 @@
 #include "DynamicArray/DynamicArray.hpp"
 #include "HelpTest/HelpTest.hpp"
 
-namespace self
-{
+namespace self {
 
 template <typename T>
-class DynamicArrayTest : public ::testing::Test
-{
+class DynamicArrayTest : public ::testing::Test {
 public:
     T obj0; T obj1; T obj2; T obj3;
     T obj_1; T obj_2; T objA;
@@ -62,7 +60,7 @@ public:
         size = 8; setIndex = 5;
         arr = new T[size];
         for (INTEGER i = 0; i < size; ++i) {
-            UDTfT obbj(i + 1, 'a' + i, double(double(i + 1) * 2.5));
+            UDTfT obbj(i + 1, 'a' + i, static_cast<double>(i + 1) * 2.5);
             da.append(obbj);
             arr[i] = obbj;
         }
@@ -73,7 +71,7 @@ public:
     ENABLE_IF(!IS_SAME(Q, char) && !IS_SAME(Q, INTEGER) && !IS_SAME(Q, UDTfT), void)
     initialize_dependent() {
         throw exception(
-            "Unrecognized type for Dynamic Array Test:" + 
+            "Unrecognized type for Dynamic Array Test:" +
             std::string(typeid(Q()).name())
         );
     }
@@ -91,31 +89,31 @@ TYPED_TEST_P(DynamicArrayTest, SizeTest) {
     EXPECT_EQ(this ->  da.size(), this -> size);
     EXPECT_EQ(this -> da0.size(), 0);
 }
-TYPED_TEST_P (DynamicArrayTest, getTest) {
+TYPED_TEST_P(DynamicArrayTest, getTest) {
     EXPECT_EQ(this -> da.get(0), this -> obj0);
     EXPECT_EQ(this -> da.get(2), this -> obj2);
     EXPECT_EQ(this -> da[0], this -> obj0);
     EXPECT_EQ(this -> da[2], this -> obj2);
 }
-TYPED_TEST_P (DynamicArrayTest, getNegativeIndexTest) {
+TYPED_TEST_P(DynamicArrayTest, getNegativeIndexTest) {
     EXPECT_EQ(this -> da.get(-1), this -> obj_1);
     EXPECT_EQ(this -> da.get(-2), this -> obj_2);
 }
-TYPED_TEST_P (DynamicArrayTest, getInvalidIndexTest) {
+TYPED_TEST_P(DynamicArrayTest, getInvalidIndexTest) {
     EXPECT_THROW(this -> da.get(this -> negativeIndex), exception);
-    EXPECT_THROW(this -> da.get(this -> largeIndex), exception);   
+    EXPECT_THROW(this -> da.get(this -> largeIndex), exception);
 }
-TYPED_TEST_P (DynamicArrayTest, setTest) {
+TYPED_TEST_P(DynamicArrayTest, setTest) {
     this -> da.set(this -> objA, this -> setIndex);
     EXPECT_EQ(this -> da.get(this -> setIndex), this -> objA);
     EXPECT_THROW(this -> da.set(this -> objA, this -> negativeIndex), exception);
     EXPECT_THROW(this -> da.set(this -> objA, this -> largeIndex), exception);
 }
-TYPED_TEST_P (DynamicArrayTest, setUsingGetTest) {
+TYPED_TEST_P(DynamicArrayTest, setUsingGetTest) {
     this -> da[this -> setIndex] = this -> objA;
     EXPECT_EQ(this -> da.get(this -> setIndex), this -> objA);
 }
-TYPED_TEST_P (DynamicArrayTest, addTest) {
+TYPED_TEST_P(DynamicArrayTest, addTest) {
     this -> da0.append(this -> objA);
     EXPECT_EQ(this -> da0.get(0), this -> objA);
     this -> da.add(this -> objA, (this -> size) - 1);
@@ -124,30 +122,30 @@ TYPED_TEST_P (DynamicArrayTest, addTest) {
     EXPECT_THROW(this -> da.add(this -> objA, this -> negativeIndex), exception);
     EXPECT_THROW(this -> da.add(this -> objA, this -> largeIndex), exception);
 }
-TYPED_TEST_P (DynamicArrayTest, findTest) {
+TYPED_TEST_P(DynamicArrayTest, findTest) {
     EXPECT_EQ(this -> da.find(this -> obj1), 01);
     EXPECT_EQ(this -> da.find(this -> objA), -1);
     EXPECT_TRUE(this -> da.contains(this -> obj2));
     EXPECT_FALSE(this -> da.contains(this -> objA));
 }
-TYPED_TEST_P (DynamicArrayTest, removeTest) {
+TYPED_TEST_P(DynamicArrayTest, removeTest) {
     EXPECT_TRUE(this -> da.contains(this -> obj3));
     EXPECT_TRUE(this -> da.remove(this -> obj3));
     EXPECT_FALSE(this -> da.contains(this -> obj3));
 }
-TYPED_TEST_P (DynamicArrayTest, iteratorTest) {
+TYPED_TEST_P(DynamicArrayTest, iteratorTest) {
     INTEGER ii = 0;
     for (TypeParam ele:this -> da) {
         ASSERT_EQ(ele, this -> arr[ii++]);
     }
 }
-TYPED_TEST_P (DynamicArrayTest, iteratorInvalidationTest) {
+TYPED_TEST_P(DynamicArrayTest, iteratorInvalidationTest) {
     typename self::DynamicArray<TypeParam>::iterator ii = this -> da.begin();
     this -> da.append(this -> obj0);
     EXPECT_THROW(ii != this -> da.end(), self::exception);
 }
 
-REGISTER_TYPED_TEST_SUITE_P( DynamicArrayTest,
+REGISTER_TYPED_TEST_SUITE_P(DynamicArrayTest,
     constructorExceptionTest,
     SizeTest,
     getTest,
